fix(building-rollup): add more examples to docs
