+++ packages/building-rollup/demo/js/.babelrc
