+++ packages/building-rollup/demo/js/a/b/foo.txt
