+++ packages/building-rollup/demo/js/a/b/meta-url-test-3.js
