+++ packages/building-rollup/demo/js/a/meta-url-test-2.js
