+++ packages/building-rollup/demo/js/demo-app.js
