+++ packages/building-rollup/demo/js/demo-component.js
