+++ packages/building-rollup/demo/js/lazy/import-meta/meta-url-test-4.js
