+++ packages/building-rollup/demo/js/lazy/lazy-component.js
