+++ packages/building-rollup/demo/ts-babel/index.html
@@ -1,0 +1,19 @@
+<!doctype html>
+<html>
+
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover" />
+  <meta http-equiv="expires" content="0">
+
+  <title>My Demo</title>
+</head>
+<body>
+
+  <h1>Static content in index.html is preserved</h1>
+
+  <demo-app></demo-app>
+  <script type="module" src="./demo-app.ts"></script>
+</body>
+
+</html>
