+++ packages/building-rollup/demo/ts-babel/rollup.config.js
@@ -1,0 +1,5 @@
+import createDefaultConfig from '../../modern-and-legacy-config';
+
+export default createDefaultConfig({
+  input: './demo/ts-babel/index.html',
+});
