+++ packages/building-rollup/package.json
@@ -14,13 +14,16 @@
     "directory": "packages/building-rollup"
   },
   "scripts": {
-    "build:demo": "rimraf dist && rollup -c demo/rollup.config.js",
+    "build:demo": "rimraf dist && rollup -c demo/js/rollup.config.js",
     "start:demo": "http-server dist -o",
-    "watch:demo": "rimraf dist && rollup -c demo/rollup.config.js --watch & http-server dist -o",
+    "watch:demo": "rimraf dist && rollup -c demo/js/rollup.config.js --watch & http-server dist -o",
     "visualizer:demo": "rimraf dist && rollup -c demo/rollup.visualizer.config.js",
-    "build-modern:demo": "rimraf dist && rollup -c demo/rollup.modern.config.js",
-    "start-modern:demo": "http-server dist -o",
-    "watch-modern:demo": "rimraf dist && rollup -c demo/rollup.modern.config.js --watch & http-server dist -o",
+    "build:modern": "rimraf dist && rollup -c demo/js/rollup.modern.config.js",
+    "watch:modern": "rimraf dist && rollup -c demo/js/rollup.modern.config.js --watch & http-server dist -o",
+    "build:ts": "rimraf dist && rollup -c demo/ts/rollup.config.js",
+    "watch:ts": "rimraf dist && rollup -c demo/ts/rollup.config.js --watch & http-server dist -o",
+    "build:ts-babel": "rimraf dist && rollup -c demo/ts-babel/rollup.config.js",
+    "watch:ts-babel": "rimraf dist && rollup -c demo/ts-babel/rollup.config.js --watch & http-server dist -o",
     "prepublishOnly": "../../scripts/insert-header.js"
   },
   "keywords": [
@@ -50,13 +53,14 @@
   "devDependencies": {
     "@babel/plugin-proposal-class-properties": "^7.3.4",
     "@babel/plugin-proposal-decorators": "^7.3.0",
+    "@babel/preset-typescript": "^7.3.3",
     "http-server": "^0.11.1",
     "lit-element": "^2.0.1",
     "owc-dev-server": "^0.2.0",
     "rimraf": "^2.6.3",
     "rollup": "^1.4.0",
-    "rollup-plugin-copier": "^1.1.0",
-    "rollup-plugin-copy-assets": "^1.1.0",
+    "rollup-plugin-cpy": "^1.1.0",
+    "rollup-plugin-typescript2": "^0.19.3",
     "rollup-plugin-visualizer": "^1.0.2"
   }
 }
